:sparkles: :: 저장된 데이터를 불러오고 보낸 후 예외처리 하는 코드 추가
