+++ presentation/src/main/java/com/sms/presentation/main/ui/fill_out_information/screen/AwardScreen.kt
@@ -1,5 +1,6 @@
 package com.sms.presentation.main.ui.fill_out_information.screen
 
+import android.util.Log
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.itemsIndexed
@@ -15,7 +16,10 @@ import com.msg.sms.design.component.spacer.SmsSpacer
 import com.sms.presentation.main.ui.fill_out_information.component.award.AwardBottomButtonComponent
 import com.sms.presentation.main.ui.fill_out_information.component.award.AwardComponent
 import com.sms.presentation.main.ui.fill_out_information.data.AwardData
+import com.sms.presentation.main.ui.util.isEmailRegularExpression
+import com.sms.presentation.main.ui.util.isUrlRegularExpression
 import com.sms.presentation.main.viewmodel.FillOutViewModel
+import com.sms.presentation.main.viewmodel.util.Event
 
 
 @Composable
@@ -26,6 +30,13 @@ fun AwardScreen(
     onPreviousButtonClick: () -> Unit,
     onDateBottomSheetOpenButtonClick: (index: Int) -> Unit
 ) {
+    val enteredProfileData = viewModel.getEnteredProfileInformation()
+    val enteredSchoolLifeData = viewModel.getEnteredSchoolLifeInformation()
+    val enteredWorkConditionData = viewModel.getEnteredWorkConditionInformation()
+    val enteredMilitaryServiceData = viewModel.getEnteredMilitaryServiceInformation()
+    val enteredCertificateData = viewModel.getEnteredCertification()
+    val enteredForeignLanguagesData = viewModel.getEnteredForeignLanguagesInformation()
+
     val awardList = remember {
         mutableStateListOf(AwardData("", "", "", isToggleOpen = true))
     }
@@ -73,5 +84,86 @@ fun AwardScreen(
                 }
             )
         }
+    }
+}
+
+suspend fun imageFileUpload(
+    viewModel: FillOutViewModel,
+    dialog: (visible: Boolean, title: String, msg: String) -> Unit,
+    isUnauthorized: () -> Unit,
+    isBadRequest: () -> Unit,
+) {
+    viewModel.imageUploadResponse.collect { response ->
+        viewModel.specifyWhenCompleteFileUpload()
+        when (response) {
+            is Event.Success -> {
+                viewModel.setProfileImageUrl(response.data!!.fileUrl)
+            }
+            is Event.Unauthorized -> {
+                dialog(true, "토큰 만료", "다시 로그인 해주세요")
+                isUnauthorized()
+            }
+            is Event.BadRequest -> {
+                dialog(true, "에러", "파일이 jpg, jpeg, png, heic 가 아닙니다.")
+                isBadRequest()
+            }
+            is Event.Loading -> {}
+            else -> {
+                dialog(true, "에러", "알 수 없는 오류 발생")
+            }
+        }
+    }
+}
+
+suspend fun enterStudentInformation(
+    viewModel: FillOutViewModel,
+    dialog: (visible: Boolean, title: String, msg: String) -> Unit,
+    onDialogButtonClick: () -> Unit,
+    isSuccess: () -> Unit,
+) {
+    viewModel.enterInformationResponse.collect { response ->
+        Log.d("정보기입", response.toString())
+        when (response) {
+            is Event.Success -> {
+                dialog(true, "성공", "정보기입을 완료했습니다.")
+                isSuccess()
+            }
+
+            is Event.BadRequest -> {
+                if (!viewModel.getEnteredProfileInformation().contactEmail.isEmailRegularExpression()) {
+                    dialog(true, "에러", "이메일 형식이 맞지 않습니다. \n수정하시겠습니까?")
+                    onDialogButtonClick()
+                } else if (!viewModel.getEnteredProfileInformation().portfolioUrl.isUrlRegularExpression()) {
+                    dialog(true, "에러", "포트폴리오 Url이 형식에 맞지 않습니다. \n수정하시겠습니까?")
+                    onDialogButtonClick()
+                } else {
+                    dialog(true, "에러", "이메일 형식또는 url형식이 맞지 않습니다. \n수정하시겠습니까?")
+                    onDialogButtonClick()
+                }
+            }
+
+            is Event.Conflict -> {
+                dialog(true, "에러", "이미 존재하는 유저 입니다.")
+            }
+
+            is Event.Loading -> {}
+            else -> {
+                dialog(true, "에러", "알 수 없는 오류 발생")
+            }
+        }
+    }
+}
+
+private fun String.toEnum(): String {
+    return when (this) {
+        "정규직" -> "FULL_TIME"
+        "비정규직" -> "TEMPORARY"
+        "계약직" -> "CONSTRACT"
+        "인턴" -> "INTERN"
+        "병특 희망" -> "HOPE"
+        "희망하지 않음" -> "NOT_HOPE"
+        "상관없음" -> "NO_MATTER"
+        "해당 사항 없음" -> "NONE"
+        else -> ""
     }
 }
